Main Final Notebook Posted
Hey, I started the full final notebook, both of you are free to change anything in it. 
Would like to determine the "main metric", I wrote some ideas down, once the writing about that is done the "The Metric" section will be done. 
Also, add your data charts and stuff for analysis somewhere under the "Best Neighborhood" section, once it is all compiled we can determine which neighborhood to choose as best. 
Conclusions can't be written until the best neighborhood is determined, so don't worry about it until then.